cygwin proofed
git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@1477 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ code/base/buildscripts/testrun_results.rb
@@ -31,16 +31,16 @@ class TestRunResults
         root_dir.ensure_directory
         
         # create symlink to latest testrun, must have Cygwin on Windows for this to work
-        begin
+        # detect if we're in unix or cygwin env
+        uname = `uname 2>&1`
+        unless $? != 0      
           unixpath = root_dir.to_s.gsub(/\\/, "/")
-          if ENV['OS'] =~ /(Windows)|(CYGWIN)/i            
+          if ENV['OS'] =~ /(Windows)|(CYGWIN)/i 
               unixpath = `cygpath -l -w #{unixpath}`.gsub(/\\/, "/")
               unixpath = `cygpath -u #{unixpath}`.strip
           end
                     
-          link=`rm testrun-latest; ln -s #{unixpath} testrun-latest`
-        rescue
-          # if fails, do nothing
+          link=`rm testrun-latest; ln -s #{unixpath} testrun-latest`        
         end
         
         TestRunResults.new(root_dir)
